changed loglevel of logger object to DEBUG always
Previous to this change the logger objects loglevel
was always set to the loglevel of the StreamHandler.
This limited the possible loglevels for logfiles
to be equal to or higher than this loglevel, since
the loglevel of handlers added to the logger object
are limited be the loglevel of the latter.

With the logger object having the lowest possible
loglevel it is now possible to define a logfile
with a lower loglevel (e.g. DEBUG) than the console
output (e.g. INFO).