Tweaked a bit to address issue #285
The formatter that have been supplied sets the formatter only for console and if it sets the same for logfile, it prints extra color character which is unwanted. 

Basically when we supply formatter, the formatter only gets applied to console not the logfile (if logfile have been provided). That makes the output inconsistent.


Ref. of Issue:
https://github.com/metachris/logzero/issues/285